requirements.txt
streamlit
openai
mem0ai
litellm
qdrant-client
